[FEAT] 20220907 오후 수업 내용

+++ django2/articles/views.py
@@ -2,6 +2,7 @@ from django.shortcuts import render, redirect
 from .models import Article
 from .forms import ArticleForm
 from django.views.decorators.http import require_safe, require_http_methods, require_POST
+from django.contrib.auth.decorators import login_required
 
 # Create your views here.
 @require_safe
@@ -19,6 +20,7 @@ def index(request):
 #     }
 #     return render(request, "articles/new.html", context)
 
+@login_required
 @require_http_methods(["GET", "POST"])
 def create(request):
     if request.method == "POST":
@@ -44,10 +46,12 @@ def detail(request, pk):
     }
     return render(request, "articles/detail.html", context)
 
+@login_required
 @require_POST
 def delete(request, pk):
-    article = Article.objects.get(pk=pk)
-    article.delete()
+    if request.user.is_authenticated:
+        article = Article.objects.get(pk=pk)
+        article.delete()
     return redirect("articles:index")
 
 # def edit(request, pk):
@@ -59,6 +63,7 @@ def delete(request, pk):
 #     }
 #     return render(request, "articles/edit.html", context)
 
+@login_required
 @require_http_methods(["GET", "POST"])
 def update(request, pk):
     article = Article.objects.get(pk=pk)
